録音下書き保存

+++ lib/presentation/ui/pages/talk/post_recording_screen.dart
@@ -10,6 +10,7 @@ import 'package:pop_talk/presentation/ui/templates/talk/after_recording_page.dar
 import 'package:pop_talk/presentation/ui/templates/talk/before_recording_page.dart';
 import 'package:pop_talk/presentation/ui/templates/talk/during_recording_page.dart';
 import 'package:pop_talk/presentation/ui/templates/talk/talk_edit_page.dart';
+import 'package:uuid/uuid.dart';
 
 enum ScreenState {
   beforeRecording,
@@ -19,8 +20,11 @@ enum ScreenState {
 }
 
 class PostRecordingScreen extends StatefulWidget {
-  const PostRecordingScreen({required this.talkTopicId});
+  const PostRecordingScreen({
+    required this.talkTopicId,
+    required this.talkTopicName});
   final String talkTopicId;
+  final String talkTopicName;
 
   @override
   _PostRecordingScreenState createState() => _PostRecordingScreenState();
@@ -37,6 +41,7 @@ class _PostRecordingScreenState extends State<PostRecordingScreen> {
   String _path = '';
   bool _isRecorderInitiated = false;
   Duration _duration = Duration();
+  bool _isSaveFile = false;
 
   @override
   void initState() {
@@ -56,9 +61,12 @@ class _PostRecordingScreenState extends State<PostRecordingScreen> {
     _stopRecording();
     _flutterSoundRecorder.closeAudioSession();
 
-    final outputFile = File(_path);
-    if (outputFile.existsSync()) {
-      outputFile.delete();
+    if(!_isSaveFile){
+      final outputFile = File(_path);
+      if (outputFile.existsSync()) {
+        outputFile.delete();
+      }
+
     }
 
     if (_recorderSubscription != null) {
@@ -69,7 +77,6 @@ class _PostRecordingScreenState extends State<PostRecordingScreen> {
 
   @override
   Widget build(BuildContext context) {
-    //todo single child scroll view for when keyboard pops up
 
     return Scaffold(
       body: Stack(alignment: AlignmentDirectional.topStart, children: [
@@ -93,24 +100,28 @@ class _PostRecordingScreenState extends State<PostRecordingScreen> {
     switch (_screenState) {
       case ScreenState.beforeRecording:
         page = BeforeRecordingPage(
-            onRecordingButtonPressed: _onRecordingButtonPressed);
+            onRecordingButtonPressed: _onRecordingButtonPressed,
+            talkTopicName: widget.talkTopicName,);
         break;
       case ScreenState.duringRecording:
         page = DuringRecordingPage(
             onStopButtonPressed: _onStopButtonPressed,
-            stream: _flutterSoundRecorder.onProgress!,);
+            stream: _flutterSoundRecorder.onProgress!,
+            talkTopicName: widget.talkTopicName,);
         break;
       case ScreenState.afterRecording:
         page = AfterRecordingPage(
             onAgainButtonPressed: _onAgainButtonPressed,
-            onEditButtonPressed: _onEditButtonPressed,);
+            onEditButtonPressed: _onEditButtonPressed,
+            talkTopicName: widget.talkTopicName,);
         break;
       case ScreenState.edit:
         page = TalkEditPage(
             titleController: _titleController,
             descriptionController: _descriptionController,
             onPostButtonPressed: _onPostButtonPressed,
-            onDraftSaveButtonPressed: _draftSaveButtonPressed,);
+            onDraftSaveButtonPressed: _draftSaveButtonPressed,
+            talkTopicName: widget.talkTopicName,);
         break;
     }
     return page;
@@ -151,6 +162,9 @@ class _PostRecordingScreenState extends State<PostRecordingScreen> {
   Future<void> _onPostButtonPressed() async{
     final recordingNotifier
                       = context.read<RecordingNotifier>(recordingProvider);
+
+    //todo prompt login/signup if not done yet
+
     await recordingNotifier.postRecording(
       title: _titleController.text,
       description: _descriptionController.text,
@@ -159,10 +173,22 @@ class _PostRecordingScreenState extends State<PostRecordingScreen> {
       talkTopicId: widget.talkTopicId
     );
 
+
     Navigator.pop(context);
   }
 
-  void _draftSaveButtonPressed() {
+  Future<void> _draftSaveButtonPressed() async{
+    final recordingNotifier
+                    = context.read<RecordingNotifier>(recordingProvider);
+    await recordingNotifier.saveDraft(
+      title: _titleController.text,
+      description: _descriptionController.text,
+      path: _path,
+      duration: _duration,
+      talkTopicId: widget.talkTopicId,
+    );
+
+    _isSaveFile = true;
     Navigator.pop(context);
 
   }
@@ -170,28 +196,22 @@ class _PostRecordingScreenState extends State<PostRecordingScreen> {
 
   //-----------------------------------------------------------Recording Methods
   Future<void> _openTheRecorder() async {
-    //upgraded the minimum SDK version of Android(23) to encode AAC ADTS
-    //todo [check] upgrade minimum OS version of iOS(10.0)?
-
     //todo permission
 
-    final temDir = await getTemporaryDirectory();
-    _path = '${temDir.path}/flutter_sound_example.aac';
-    final outputFile = File(_path);
-    if (outputFile.existsSync()) {
-      await outputFile.delete();
-    }
+    final docDir = await getApplicationDocumentsDirectory();
+    final fileName = const Uuid().v1();
+    _path = '${docDir.path}/$fileName.aac';
+
     await _flutterSoundRecorder.openAudioSession();
   }
 
   Future<void> _setSubscriptionDuration() async {
-    //todo [check] can be less than 1 second?
     await _flutterSoundRecorder
-        .setSubscriptionDuration(const Duration(seconds: 1));
+        .setSubscriptionDuration(const Duration(milliseconds: 1));
   }
 
   Future<void> _startRecording() async {
-    if (!_isRecorderInitiated) return; //todo [check] maybe assert?
+    assert(_isRecorderInitiated);
     await _flutterSoundRecorder.startRecorder(
       toFile: _path,
       codec: Codec.aacADTS,
